Junit demo - test case fail will force us to look at our implementation

+++ PHASE4/junit-demo/src/main/java/com/sl/Calculator.java
@@ -9,7 +9,7 @@ public class Calculator {
 	
 	
 	public float divide(int a, int b){		
-		return a/b;		
+		return (float) a/ (float) b;		
 	}
